requirements.txt
numpy==1.20.2
scipy==1.6.3
pandas==1.2.4
dask>=2020.12,<=2021.5.0
joblib==1.0.1
seaborn==0.11.1
bokeh==2.2.3
pytest-randomly==3.5.0
factory-boy==3.2.0
astropy==4.2
aplpy==2.0.3
reproject==0.7.1
avro-python3==1.10.2
fastavro==1.2.4
tqdm>=4.23,<4.61
matplotlib==3.4.1
astroquery==0.4.1
apispec==4.0.0
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.3
marshmallow-enum==1.5.1
Pillow==8.2.0
sncosmo==2.5.0
tdtax==0.1.3
healpix-alchemy==0.1.2
jsonschema==3.2.0
jsonpath_ng==1.5.2
pytest-rerunfailures==9.1.1
timezonefinder==5.2.0
astroplan==0.8
py3-validate-email==0.2.12
alembic==1.5.5
marshmallow-jsonschema==0.11.1
requests_toolbelt==0.9.1
sendgrid==6.7.0
twilio==6.50.1
jsonschema-to-openapi==0.2.1
responses==0.13.3
numba==0.53.1
pyvo==1.1
lxml==4.6.3
suds-py3==1.4.4.1
vcrpy==4.1.1
aiosmtpd==1.4.2
